requirements.txt
annotated-types==0.7.0
argcomplete==3.5.1
click==8.1.7
iniconfig==2.0.0
openapi-generator-cli==7.10.0.post0
packaging==24.2
pipx==1.7.1
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
setuptools==75.6.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.2.3
userpath==1.9.2
